+++ src/lib.rs
@@ -6,9 +6,12 @@ extern crate serde_json;
 extern crate socool_parser;
 #[macro_use]
 extern crate serde_derive;
+extern crate difference;
+extern crate indexmap;
 extern crate num_rational;
 extern crate rand;
 extern crate rayon;
+extern crate term;
 
 #[macro_use]
 pub mod macros;
@@ -22,5 +25,6 @@ pub mod portaudio_setup;
 pub mod render;
 pub mod ring_buffer;
 pub mod settings;
+pub mod testing;
 pub mod ui;
 pub mod write;
